got this working for my purposes, could be more polished, but calling it verion 0.2
This allows you to run an instance as the "master" and multiple others as slaves.  Concept is that the master will handle talking to redis and converting master/slave of redis when needed

the slaves will ping the master at the specified interval, for the purpose of updating nginx config files.

this allows us to connect to the same redis instance from multiple ip address at the load balancer side, which finally gets us true redundant availability